requirements/nis_requirements.txt
absl-py==1.2.0
addict==2.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
basicsr==1.4.2
bleach==4.1.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==4.4.2
deepspeed==0.7.7
defusedxml==0.7.1
einops==0.4.1
entrypoints==0.4
future==0.18.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
hjson==3.1.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jupyterthemes==0.20.0
kiwisolver==1.3.1
kornia==0.6.6
lesscpy==0.15.0
lmdb==1.4.1
lpips==0.1.4
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
ninja==1.11.1
notebook==6.4.9
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
oauthlib==3.2.0
olefile==0.46
omnicv==1.1.3
opencv-contrib-python==4.7.0.68
opencv-python==4.5.5.64
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
ply==3.11
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.4
ptflops==0.6.9
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.0
pysrwarp==1.0.0
python-dateutil==2.8.2
pytorch-msssim==0.2.1
pytorch-ssim==0.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
tb-nightly==2.11.0a20220816
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
terminado==0.12.1
testpath==0.6.0
tifffile==2020.9.3
timm==0.5.4
torch==1.10.2
torchaudio==0.10.2
torchgeometry==0.1.2
torchvision==0.11.3
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
yapf==0.32.0
zipp==3.6.0
